renamed mod and moved constants
