+++ sudo_plugin/src/plugin/mod.rs
@@ -20,7 +20,7 @@ mod command_info;
 mod settings;
 mod user_info;
 mod print_facility;
-mod conv_prompt;
+mod conv_facility;
 mod traits;
 
 use super::errors::*;
@@ -28,7 +28,7 @@ use super::version::Version;
 
 pub use self::option_map::OptionMap;
 pub use self::print_facility::PrintFacility;
-pub use self::conv_prompt::ConversationPrompt;
+pub use self::conv_facility::ConversationPrompt;
 
 use self::command_info::CommandInfo;
 use self::settings::Settings;
